Rename RecipeParser -> Parser, Parser -> Lexer, remove unused files
